+++ src/main/java/dao/UserDao.java
@@ -5,9 +5,9 @@ import spring.model.User;
 public interface UserDao {
     void create(User user);
 
-    User read(int id);
+    User read(Long id);
 
     void update(User user);
 
-    void delete(int id);
+    void delete(long id);
 }
